requirements.txt
Faker==37.6.0
psycopg2==2.9.6
python-dotenv
